+++ taktik/core/social_media/instagram/actions/atomic/navigation/search_navigation.py
@@ -6,7 +6,8 @@ from typing import Optional
 from loguru import logger
 
 from ...core.base_action import BaseAction
-from ....ui.selectors import DETECTION_SELECTORS, NAVIGATION_SELECTORS, PROFILE_SELECTORS
+from ....ui.selectors.shell.navigation import NAVIGATION_SELECTORS
+from ....ui.selectors.surfaces.profile import PROFILE_SELECTORS
 from taktik.core.clone import get_active_package